+++ src/main/java/devocean/tickit/dto/test/TestSignInRequestDto.java
@@ -1,8 +1,9 @@
-package devocean.tickit.dto;
+package devocean.tickit.dto.test;
 
 import jakarta.validation.constraints.NotNull;
 
-public record TestLogInRequestDto(
+public record TestSignInRequestDto(
+        @NotNull String name,
         @NotNull String provider,
         @NotNull String providerId
 ) {
