+++ src/main/java/devocean/tickit/dto/user/UserDto.java
@@ -1,4 +1,4 @@
-package devocean.tickit.dto;
+package devocean.tickit.dto.user;
 
 import devocean.tickit.global.constant.Role;
 import jakarta.validation.constraints.NotNull;
