+++ src/main/java/devocean/tickit/service/AuthenticationService.java
@@ -6,9 +6,9 @@ import com.google.api.client.http.javanet.NetHttpTransport;
 import com.google.api.client.json.jackson2.JacksonFactory;
 import devocean.tickit.domain.User;
 import devocean.tickit.domain.UserInfo;
-import devocean.tickit.dto.AuthenticationResponseDto;
-import devocean.tickit.dto.AuthenticationRequestDto;
-import devocean.tickit.dto.UserDto;
+import devocean.tickit.dto.auth.AuthenticationRequestDto;
+import devocean.tickit.dto.auth.AuthenticationResponseDto;
+import devocean.tickit.dto.user.UserDto;
 import devocean.tickit.global.api.ApiResponse;
 import devocean.tickit.global.api.ErrorCode;
 import devocean.tickit.global.api.UnauthorizedException;
